remove unnecessary validator tests
